// src/enum/CommonCityId.enum.ts
export enum CommonCityIdEnum {
    vancouver = 1, // these were decided by the way /seed/gyms/* came out.
    calgary = 2, // the #s are manually transcribed from those files.
    edmonton = 3,
    winninpeg = 4,
    toronto = 5,
    montreal = 6,
    mississauga = 7,
    brampton = 8,
    hamilton = 9,
    ottawa = 10,
}